+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Adds `Overstrike.compose`.
+
 ## 0.2.0 (Aug 17, 2014)
 - Adds `Overstrike.new`.
 - Adds `Overstrike.noop`.